Extracted drawing code into seperate files
